Changed cash to double instead of int
Changed all the monetary values to double
so cents can also be displayed.

+++ lib/entities/cash.dart
@@ -1,4 +1,6 @@
+import 'dart:ffi';
+
 class Cash {
-  Map<String, int> cashPerDay;
+  Map<String, double> cashPerDay;
   Cash({required this.cashPerDay});
 }
